+++ CHANGELOG.md
@@ -4,6 +4,8 @@ ChangeLog
 * Compatibility improvements with docker-py mainline and Docker versions
 * Parallelization of orchestration plays, respecting dependency and
   execution order but doing as much work as possible in parallel
+* Dropped support for `docker_endpoint` in ship configuration
+* SSH tunneling support via `bgtunnel` on-demand SSH tunnels
 
 Maestro 0.1.8.1
 ---------------